[fix] pages 폴더안의 파일들 각자 폴더안으로 이동
