+++ src/Router.tsx
@@ -6,20 +6,19 @@ import KakaoRedirection from './pages/signup/KakaoRedirection'
 import SignUpTerms from './pages/signup/SignUpTerms'
 import SignUpNickname from './pages/signup/SignUpNickname'
 import Setting from './pages/setting/Setting'
-import EditProfile from './pages/EditProfile'
+import EditProfile from './pages/setting/EditProfile'
 import SignUpOnBoarding from './pages/signup/SignUpOnBoarding'
-import Answer from './pages/Answer'
-import Folder from './pages/Folder'
-import AnswerComplete from './pages/AnswerComplete'
-import Main from './pages/Main'
-import QuestionList from './pages/QuestionList'
-import GroupPlus from './pages/GroupPlus'
-import GroupModify from './pages/GroupModify'
+import Answer from './pages/answer/Answer'
+import Folder from './pages/answer/Folder'
+import AnswerComplete from './pages/answer/AnswerComplete'
+import Main from './pages/main/Main'
+import QuestionList from './pages/question/QuestionList'
+import GroupPlus from './pages/category/GroupPlus'
+import GroupModify from './pages/category/GroupModify'
 import PrivacyPolicy from './pages/setting/PrivacyPolicy'
-import Ask from './pages/Ask'
+import Ask from './pages/setting/Complaint'
 import Term from './pages/setting/Tos'
 import DetailTestPage from './pages/DetailTestPage'
-
 
 const Router = () => {
   return (
