+++ src/pages/answer/Answer.tsx
@@ -1,13 +1,13 @@
 import styled from 'styled-components'
 import { useState } from 'react'
-import { colors } from '../styles/colors'
-import Question from '../components/answer/Question'
-import { UnFixedButton } from '../components/common/Button'
-import Music from '../components/answer/Music'
-import Link from '../components/answer/Link'
-import DelayModal from '../components/common/DelayModal'
-import Modal from '../components/common/Modal'
-import PersonalHeader from '../components/answer/PersonalHeader'
+import { colors } from '../../styles/colors'
+import Question from '../../components/answer/Question'
+import { UnFixedButton } from '../../components/common/Button'
+import Music from '../../components/answer/Music'
+import Link from '../../components/answer/Link'
+import DelayModal from '../../components/common/DelayModal'
+import Modal from '../../components/common/Modal'
+import PersonalHeader from '../../components/answer/PersonalHeader'
 import { useNavigate } from 'react-router-dom'
 
 const Answer = () => {
